+++ app/GraphQL/Mutations/UpdateProperty.php
@@ -17,6 +17,7 @@ use Illuminate\Support\Facades\Storage;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use App\Exceptions\SendException;
 use App\Http\Traits\GetIdTrait;
+use Image;
 
 
 class UpdateProperty
@@ -140,6 +141,15 @@ class UpdateProperty
           };
           $property_image->storeAs('public/property',$fileName_img);
                if(file_exists(storage_path('app/public/property/'.$property_image))){
+
+                $image = Image::make(storage_path('app/public/property/'.$fileName_img));
+
+                $image->resize(null, 200, function($constraint) {
+                    $constraint->aspectRatio();
+                });
+        
+                $image->save(storage_path('app/public/property/min/'.$fileName_img));
+
                    PropertyImage::create([
                        'property_id' => $property_id,
                        'name'        => $fileName_img,
@@ -157,11 +167,14 @@ class UpdateProperty
 
      foreach($property_images_delete_ids as $images_id){
         $propertyImage=PropertyImage::find($images_id);
-        if($user_auth->id == $propertyImage->property->user_id){
+        if($propertyImage&&$user_auth->id == $propertyImage->property->user_id){
             $propertyImage_name = $propertyImage->getOriginal('name');
             if($propertyImage_name&&file_exists(storage_path('app/public/property/'.$propertyImage_name))){
                 unlink(storage_path('app/public/property/'. $propertyImage_name));
-              }
+            }
+            if($propertyImage_name&&file_exists(storage_path('app/public/property/min/'.$propertyImage_name))){
+                unlink(storage_path('app/public/property/min/'.$propertyImage_name));
+            }   
               $propertyImage->delete();
         }
